subir

+++ euc05-sitio-solicitudes-inversiones/euc05-service/src/main/java/com/citi/euces/solicitudes/entities/AutoTasasPorEjecutivo.java
@@ -38,7 +38,7 @@ public class AutoTasasPorEjecutivo {
 	@Column(name ="NOMINA")
 	private String tas_NOMINA;
 	@Column(name ="MONTO")
-	private BigInteger tas_MONTO;
+	private double tas_MONTO;
 	@Column(name ="PLAZO")
 	private Long tas_PLAZO;
 	@Column(name ="TASA_AUTORI")
@@ -190,11 +190,11 @@ public class AutoTasasPorEjecutivo {
 		this.tas_NOMINA = tas_NOMINA;
 	}
 
-	public BigInteger getTas_MONTO() {
+	public double getTas_MONTO() {
 		return tas_MONTO;
 	}
 
-	public void setTas_MONTO(BigInteger tas_MONTO) {
+	public void setTas_MONTO(double tas_MONTO) {
 		this.tas_MONTO = tas_MONTO;
 	}
